Pages/profile_page.py
"""

"""

# Import libraries
import pandas as pd
import os
import streamlit as st

# Define function to show user profile
def show_user_profile(user):
    st.title("👤 Profile")
    st.image(user.get("picture"), width=150)
    st.write(f"**Name:** {user.get('name')}")
    st.write(f"**Email:** {user.get('email')}")
    st.write("Here you can view and edit your profile information.")
    st.write("More profile management features coming soon!")
    st.markdown("### 📝 Update Preferences (coming soon!)")
    st.markdown("---")
    st.info("Profile management features will be added in future updates.")
    st.markdown("Use the sidebar to navigate to other sections of the app.")
    st.markdown("For more information, visit our [GitHub repository](https://github.com/your-repo-link).")

