requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit-oauth
google-cloud-aiplatform google-auth